requirements.txt
# Requirements file for TDC Application
# Add any missing dependencies here

Django>=6.0.1
djangorestframework>=3.15.2
django-cors-headers>=4.6.0
djoser>=2.3.0
python-decouple>=3.8
pyotp>=2.9.0
qrcode>=8.0
Pillow>=11.0.0
django-filter>=24.3
gunicorn>=23.0.0
requests>=2.32.3
bleach>=6.1.0
beautifulsoup4>=4.12.0
drf-nested-routers>=0.94.1
